+++ components/ui/index.ts
@@ -1,3 +1,4 @@
 export { Navbar } from "./Navbar";
 export { Footer } from "./Footer";
 export { ThemeSwitcher } from "./ThemeSwitcher";
+export { Pagination } from "./Pagination";